+++ main/ad2_cli_cmd.h
@@ -31,6 +31,7 @@
 #define AD2_CMD_CODE     "code"
 #define AD2_CMD_PART     "partition"
 #define AD2_CMD_SOURCE   "ad2source"
+#define AD2_CMD_CONFIG   "ad2config"
 #define AD2_CMD_TERM     "ad2term"
 #define AD2_CMD_LOGMODE  "logmode"
 #define AD2_CMD_FACTORY "factory-reset"